add overrideTS to marshalOptions
